feat: adds navbar and footer components
