+++ branchout/package.json
@@ -15,7 +15,7 @@
     "@radix-ui/react-select": "^2.2.4",
     "class-variance-authority": "^0.7.1",
     "date-fns": "^4.1.0",
-    "firebase": "^11.7.3",
+    "firebase": "^11.8.0",
     "firebase-tools": "^14.3.1",
     "lucide-react": "^0.511.0",
     "next": "15.3.2",
@@ -32,14 +32,15 @@
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@types/react": "^19.1.5",
+    "@types/react-dom": "^19.1.5",
+    "@types/react-router-dom": "^5.3.3",
     "@types/tailwindcss": "^3.0.11",
     "autoprefixer": "^10.4.21",
     "eslint": "^9",
     "eslint-config-next": "15.3.2",
     "postcss": "^8.5.3",
     "tailwindcss": "^4.1.7",
-    "typescript": "5.8.3"
+    "typescript": "^5.8.3"
   }
 }
