+++ branchout/src/app/layout.tsx
@@ -1,18 +1,52 @@
-import "./globals.css"; // or wherever you keep 
+import "./globals.css"
+import type { Metadata } from "next"
+import { Inter } from "next/font/google"
+import Navbar from "@/components/navbar"
+import { GiTreeBranch } from "react-icons/gi"
+import Link from "next/link"
 
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
+const inter = Inter({ subsets: ["latin"] })
+
+export const metadata: Metadata = {
+  title: "BranchOut - Track Your Impact",
+  description: "Map, track, and share your contribution to a healthier planet.",
 }
 
 export default function RootLayout({
   children,
-}: {
+}: Readonly<{
   children: React.ReactNode
-}) {
+}>) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body className={inter.className}>
+        <div className="min-h-screen flex flex-col">
+          <Navbar />
+          <main className="flex-1 overflow-auto">
+            {children}
+          </main>
+          <footer className="border-t bg-white mt-auto">
+            <div className="container mx-auto px-4 flex flex-col gap-4 py-10 md:flex-row md:items-center md:justify-between md:py-12">
+              <div className="flex items-center gap-2">
+                <GiTreeBranch className="h-5 w-5 text-green-600" />
+                <span className="text-xl font-bold">BranchOut</span>
+              </div>
+              <p className="text-sm text-gray-500">© 2025 BranchOut. All rights reserved.</p>
+              <div className="flex gap-4">
+                <Link href="#" className="text-sm font-medium hover:text-green-600 transition-colors">
+                  Privacy
+                </Link>
+                <Link href="#" className="text-sm font-medium hover:text-green-600 transition-colors">
+                  Terms
+                </Link>
+                <Link href="#" className="text-sm font-medium hover:text-green-600 transition-colors">
+                  Contact
+                </Link>
+              </div>
+            </div>
+          </footer>
+        </div>
+      </body>
     </html>
   )
 }
