--- branchout/src/app/tailwind.config.ts
@@ -1,8 +1,0 @@
-module.exports = {
-    content: ["./src/**/*.{js,ts,jsx,tsx}"], // Make sure this matches your file paths
-    theme: {
-      extend: {},
-    },
-    plugins: [],
-  }
-
