+++ branchout/src/components/tree-details.tsx
